refactor: atualizar importações e caminhos de utilitários para o novo estado do projeto
